requirements.txt
botogram2
redis
pandas
datetime
json